% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TMEtyper.R
\name{TMEtyper}
\alias{TMEtyper}
\title{Title}
\usage{
TMEtyper(file_name, data_type)
}
\arguments{
\item{data_type}{microarray or rnaseq}
}
\value{
sample TME subtype
}
\description{
Title
}
\examples{
TMEtyper("example.txt","microarray")
}
